Реструктуризация кода и исправление зависимостей.
