+++ main.py
@@ -1,11 +1,5 @@
-from tkinter import *
-from tkinter.ttk import Treeview
-import pandas as pd
-import time
-import os.path
-import matplotlib.pyplot as plt
-from openpyxl import load_workbook
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
+from App import App
+from StatementData import StatementData
 
 """
 Заранее определим некоторые переменные
@@ -26,395 +20,18 @@ click_color = "#00a4ef"
 #  3 - Неудовлетворительное
 #  4 - Полная удовлетворённость
 pie_colors = ["#737373", "#FFB900", "#F25022", "#7FBA00"]
-    
+
 #  Путь к фоновому изображению
 background_image = "images\\background_image_radar.png"
 
 
-class App(Tk):
-    """
-    Класс графического приложения
-    """
-    
-    #  Элементы графического окна приложения
-    
-    #  Круговая диаграмма
-    fig_1 = None
-    ax_1 = None
-    patches_pie = None
-    canvas_pie = None
-    
-    #  Столбчатая диаграмма
-    fig_2 = None
-    ax_2 = None
-    patches_bar = None
-    
-    # Список обращений с неудовлетворённой тональностью
-    sd_listbox = None
-    bad_app_frame = None
-    
-    #  Таблица с данными об обращениях
-    tree = None
-    
-    # Фоновое изображение
-    bg_label = None
-    bg_image = None
-    
-    def __init__(self, data):
-        super().__init__()
-        self.update_window(data)
-        
-    def update_window(self, data):
-        """
-        Метод обновления основного окна приложения
-        """
-        
-        #  Данный участок кода будет работать при повторном обновлении данных
-        if self.fig_1 is not None:
-            #  Очистка всех графических элементов от старых значений
-            self.fig_1.clf()
-            self.fig_2.clf()
-            sd_listbox = self.sd_listbox
-            sd_listbox.delete(0, END)
-            #self.tree.delete(*tree.get_children())
-        
-        #  Настройка фонового изображения главного окна
-        bg_image = self.bg_image
-        
-        try:
-            bg_image = PhotoImage(file=background_image)
-            
-            bg_label = self.bg_label
-            bg_label = Label(self, image=bg_image)
-            bg_label.place(x=0, y=0)
-            self.bg_label = bg_label
-        except TclError:
-            print("Не могу найти изображение для фона")
-        
-        self.bg_image = bg_image
-        
-        self.title('Radar')
-        
-        ########################
-        #  Круговая диаграмма  #
-        ########################
-        
-        plt.rcParams.update({
-            "figure.facecolor":  (0.0, 0.0, 0.0, 1.0),  # red   with alpha = 30%
-            "axes.facecolor":    (0.0, 0.0, 0.0, 1.0),  # green with alpha = 50%
-            "savefig.facecolor": (0.0, 0.0, 0.0, 1.0),  # blue  with alpha = 20%
-        })
-        
-        #  Создание объекта типа figure
-        #  figure - контейнер самого верхнего уровня 
-        fig_1 = plt.Figure(figsize=(5, 5))
-        
-        #  Создание сетки для построения диаграммы 1х1 для 1 графика.
-        #  axes - область, где отображается диаграмма
-        ax_1 = fig_1.add_subplot(111)  
-        
-        #  Объект FigureCanvasTkAgg объединяет объект figure и 
-        #  графическое окно Tkinter (объект Canvas в Tkinter)
-        canvas_pie = FigureCanvasTkAgg(fig_1, self)
-        
-        #  Размещение элемента в графическом окне через метод place()
-        #  place() - абсолютное позиционирование
-        canvas_pie.get_tk_widget().place(x=30, y=50)
-        
-        #  Создаём круговую диаграмму в области axes
-        #  patches - хранит клиновидные фрагменты диаграммы
-        patches, texts, autotexts = ax_1.pie(data.summary_requester["Количество"], startangle=90, 
-            autopct='%1.1f%%', colors=pie_colors)
-        ax_1.set_title("Тональность заявителя")
-        
-        for p in patches:
-            p.set_gid(p.get_facecolor())
-            # Активируем выделение
-            p.set_picker(True)
-        
-        self.patches_pie = patches
-        
-        fig_1.canvas.mpl_connect('pick_event', self.onclick)
-        
-        self.ax_1 = ax_1
-        self.fig_1 = fig_1
-        
-        #########################
-        #  Столбчатая диаграмма #
-        #########################
-        
-        #  Cоздание фигуры
-        fig_2 = plt.Figure(figsize=(5, 5.5))
-        
-        #  Создание Axes
-        ax_2 = fig_2.add_subplot(111)
-        
-        #  Объединяем фигуру и окно программы
-        bar_canvas = FigureCanvasTkAgg(fig_2, self)
-        
-        #  Задаём позицию элемента
-        bar_canvas.get_tk_widget().place(x=460, y=50)
-        
-        #  Построение столбчатой диаграммы
-        data.summary_performer.plot(kind='bar', ax=ax_2, subplots=False, rot=0, color=['#5cb85c', '#d9534f'], width=0.08)
-        ax_2.set_title("Тональность исполнителя")
-        #ax_2.legend(fancybox=True, framealpha=0.4, shadow=True, borderpad=1)
-        
-        self.fig_2 = fig_2
-        self.ax_2 = ax_2
-        
-        """
-        #########################################################
-        #  Список обращений с неудовлетворительной тональностью #
-        #########################################################
-        
-        #  Элемент ListBox
-        bad_app_frame = tk.LabelFrame(self, text='Неудовл. исполнителя')
-        bad_app_frame.place(x=800, y=10)
-        sd_listbox = tk.Listbox(bad_app_frame, height=10)
-        sd_listbox.grid(row=1, column=1)
-        
-        #  Вставка номеров обращений с неудовл. тональностью
-        #  в список в окне приложения
-        for application_number in data.bad_app_list:
-            sd_listbox.insert(tk.END, application_number)
-        
-        self.bad_app_frame = bad_app_frame
-        self.sd_listbox = sd_listbox
-        """
-        
-        #  Запоминаем дату изменения файла
-        data.last_modified_time = time.ctime(os.path.getmtime(data.file_with_data))
-        
-        self.refresh(data)
-        
-        
-    def refresh(self, data):
-        
-        """
-        Метод обновления приложения
-        """
-        
-        try:
-            #  Запоминаем время редактирования файла с данными
-            modified_time = time.ctime(os.path.getmtime(data.file_with_data))
-            
-            #  Если она изменилась, то обновлением данные
-            if data.last_modified_time != modified_time:
-                
-                print("Обновляю данные.")
-                data.get_summary_info()
-                data.last_modified_time = modified_time
-                self.update_window(data)
-                print("Построил новые диаграммы.")
-            else:
-                print("Файл не изменён. Перерасчёт не требуется.")
-            
-        except FileExistsError:
-            print("Файл используется.")
-        
-        self.update()
-        self.after(5000, self.refresh, data)
-        
-    def onclick(self, event):
-        """
-        Действия при нажатии мышкой на фрагмент диаграммы
-        """
-        for p in self.patches_pie:
-            p.set_facecolor(p.get_gid())
-        a = event.artist
-        # previous_color = a.get_facecolor()
-        # a.set_facecolor(click_color)
-        # self.fig_1.canvas.draw()
-        # print('on pick:', a, a.get_gid())
-        
-        print("Сработало событие выделения фрагмента пирога")
-        self.show_table()
-        
-        
-    def show_table(self):
-        """
-        Метод для вывода окна с табличной информацией 
-        """
-        
-        def close_window ():
-            """
-            Метод закрытия окна с таблицей
-            """
-            window.destroy()
-            
-        window = Toplevel(self)
-        frame = Frame(window)
-        
-        tree = Treeview(window, column=("Номер обращения", "Исполнитель", "Инициатор", "Email"), show='headings', height=10)
-        tree.delete(*tree.get_children())
-        tree.column("# 1")
-        tree.heading("# 1", text="Номер обращения")
-        tree.column("# 2")
-        tree.heading("# 2", text="Исполнитель")
-        tree.column("# 3")
-        tree.heading("# 3", text="Инициатор")
-        tree.column("# 4")
-        tree.heading("# 4", text="Email")
-
-        # ~ # Вставка данных в таблицу
-        # ~ all_bad_app = data.info_bad_app
-        
-        # ~ for row in all_bad_app:
-            # ~ tree.insert('', 'end', text="1", values=(row[0], row[1], row[2], row[3]))
-        
-        #  Добавление кнопки выхода
-        button = Button(frame, text = "Закрыть", command = close_window)
-        
-        tree.pack()
-        frame.pack()
-        button.pack()
-                
-
-class Statement_Data():
-    """
-    Класс для датафрейма и оперирования над данными
-    """
-    
-    file_with_data = ""        #  Имя файла с данными
-    last_modified_time = None  #  Дата изменения файла с данными
-    dataframe = None           #  Весь датафрейм - таблица Excel-файла
-    
-    summary_performer = []     #  Датафрейм сводной таблицы по тональности заявителя
-    quantity_performer = []    #  Данные сводной таблицы по тональности исполнителя
-    labels_performer = []      #  Заголовки сводной таблицы по тональности исполнителя
-    
-    summary_requester = []     #  Датафрейм сводной таблицы по тональности заявителя
-    quantity_requester = []    #  Данные сводной по тональности заявителя
-    labels_requester = []      #  Заголовки сводной таблицы по тональности исполнителя
-    
-    all_application_file = ""  #  Файл с информацией по всем обращениям
-    bad_app_list = []          #  Список последних 10 обращений
-    #                             с неудовлетворительной тональностью исполнителя
-    
-    all_app_data = None        #  Датафрейм со всеми обращениями
-    info_bad_app = []          #  Список с информацией о всех обращениях
-    #                             с невежливым обращением.
-    
-    
-    def __init__(self, filename):
-        self.file_with_data = filename
-        self.all_application_file = all_application_file
-        self.get_summary_info()
-    
-    
-    def get_summary_info(self):
-        """ 
-        Получаем сводные таблицы
-        """
-        try:
-            self.dataframe = self.get_data_from_file(self.file_with_data) #  Подгрузим данные
-        except(FileNotFoundError):
-            print("Не могу найти файл: " + file_with_data)
-            exit(1)
-            
-        self.dataframe = self.clear_data(self.dataframe)
-        
-        #  Получаем последние 10 обращений с неудовлетворённой 
-        #  тональностью заявителя
-        bad_application = self.dataframe.loc[self.dataframe['Тональность исполнителя'] == 'Невежливое обращение']
-        bad_application = bad_application.tail(10)
-        bad_app_list = bad_application['ID Обращения'].tolist()
-        self.bad_app_list = bad_app_list
-            
-        self.summary_performer = pd.pivot_table(self.dataframe, 
-            columns="Тональность исполнителя", 
-            values="Количество", 
-            aggfunc="sum")
-        
-        self.summary_requester = pd.pivot_table(self.dataframe, 
-            index="Тональность заявителя", 
-            aggfunc="sum")
-            
-        self.all_app_data = self.get_data_from_file(self.all_application_file)
-        self.all_app_data = self.clear_data(self.all_app_data)
-        self.info_bad_app = self.get_all_bad_app(self.all_app_data, bad_app_list)
-        
-        
-    def get_data_from_file(self, name_of_file):
-        """
-        Функция чтения данных из Excel-файла
-        """
-        
-        #  Установим счётчик времени
-        start_time = time.time()
-        wb = load_workbook(filename=name_of_file, read_only=True)
-        ws = wb.active
-
-        #  Превращаем данные в Dataframe
-        data = pd.DataFrame(data=ws.values)
-
-        #  Закрытие книги
-        wb.close()
-
-        print("Загружены даные из файла ", name_of_file)
-        print("Время выполнения: ", time.time() - start_time)
-        
-        return data
-        
-        
-    def clear_data(self, data):
-        """ 
-        Функция предварительной очистки данных
-        """
-        
-        # На случай, если понадобится удалить лишние столбцы
-        #data.drop(['Решение', 'Комментарий заявителя'], axis=1, inplace=True)
-        
-        data.columns = data.iloc[0] #            Устанавливаем заголовок столбцов таблицы - значения первой строки
-        data.drop(0, axis=0, inplace=True) #     А первую строку удаляем, т.к. там по-прежнему заголовки столбцов
-        
-        data['Количество'] = 1 #                 Добавляем справа столбец "Количество", 
-        #                                        в котором все ячейки содерат значение 1 
-        #                                        (необходимо для расчёта сводной таблицы)
-        
-        return data
-        
-
-    def convert_data_for_diagrams(self, summary_table):
-        """ 
-        Функция для преобразования данных сводной таблицы
-        в данные для построения графиков
-        """
-        
-        #  Получаем названия полей
-        labels = []
-        for label in summary_table.index:
-            labels.append(label)
-            
-        #  Получаем массив значений второго столбца
-        quantity = summary_table['Количество'].tolist()
-            
-        return labels, quantity
-        
-        
-    def get_all_bad_app(self, data, list_with_app):
-        
-        all_app = []
-        
-        for one_app in list_with_app:
-            try:
-                row = data.loc[data['Номер обращения'] == one_app].values.flatten().tolist()
-                if row != []:
-                    all_app.append(row)
-            except KeyError:
-                print("Не могу найти такой номер обращения: " + one_app)
-        
-        return all_app
-
-
 if __name__ == "__main__":
     print("Подключил необходимые библиотеки")
-    
-    my_data = Statement_Data(file_with_data)
-    
-    app = App(my_data)
-    
+
+    my_data = StatementData(file_with_data, all_application_file)
+
+    app = App(my_data, background_image)
+
     #  Задаём размер окна. Сначала ширина, затем высота.
     app.geometry("1280x720")
     app.maxsize(1280, 720)
